#1334 cleanup textbox
