+++ lib/elementWrapper/elementWrapperFactory.js
@@ -18,8 +18,8 @@ const createElementWrapper = (elementType, query, attrValuePairs, _options, ...a
       return new FileFieldWrapper(selector, options, description);
     case 'Button':
       return new ButtonWrapper(selector, options, description);
-    // case 'textBox':
-    //   return new TextBoxWrapper(get, description, getIfExists);
+    case 'TextBox':
+      return new TextBoxWrapper(selector, options, description);
     // case 'radioButton':
     //   return new RadioButtonWrapper(get, description, getIfExists);
     case 'Checkbox':
