+++ lib/taiko.js
@@ -2552,67 +2552,7 @@ module.exports.textBox = textBox;
 
 function textBox(attrValuePairs, _options, ...args) {
   validate();
-  const { selector, options } = prepareParameters(attrValuePairs, _options, ...args);
-  let get;
-  const inputTypesExpression =
-    'input[type="text" i],input[type="password" i],input[type="search" i],input[type="number" i],input[type="email" i],input[type="tel" i],input[type="url" i],input:not([type]),textarea,*[contenteditable="true"]';
-  function getTextBoxElementWithLabel(label) {
-    const textBoxes = [];
-    const inputTypes = ['text', 'password', 'search', 'number', 'email', 'tel', 'url', undefined];
-    function checkAndPushElement(elem) {
-      if (
-        (elem.tagName && elem.tagName.toLowerCase() === 'textarea') ||
-        (elem.getAttribute && elem.getAttribute('contenteditable')) ||
-        (elem.tagName &&
-          elem.tagName.toLowerCase() == 'input' &&
-          elem.type &&
-          inputTypes.includes(elem.type.toLowerCase()))
-      ) {
-        textBoxes.push(elem);
-      }
-    }
-    const matchingLabels = [...document.querySelectorAll('label')].filter((labelElem) => {
-      return labelElem.innerText.toLowerCase().includes(label.toLowerCase());
-    });
-    for (let matchingLabel of matchingLabels) {
-      const labelFor = matchingLabel.getAttribute('for');
-      if (labelFor) {
-        //check label with attribute for
-        const labelForElement = document.getElementById(labelFor);
-        checkAndPushElement(labelForElement);
-      } else {
-        // check child node of label tag
-        matchingLabel.childNodes.forEach((elem) => {
-          checkAndPushElement(elem);
-        });
-      }
-    }
-    //check label is inlined
-    const inputRadioElements = [
-      ...document.querySelectorAll(
-        'input[type="text" i],input[type="password" i],input[type="search" i],input[type="number" i],input[type="email" i],input[type="tel" i],input[type="url" i],input:not([type]),textarea,*[contenteditable="true"]',
-      ),
-    ];
-    for (let inputRadioElement of inputRadioElements) {
-      if (
-        inputRadioElement.nextSibling &&
-        inputRadioElement.nextSibling.nodeType === Node.TEXT_NODE &&
-        inputRadioElement.nextSibling.wholeText.includes(label)
-      ) {
-        textBoxes.push(inputRadioElement);
-      }
-    }
-    return textBoxes;
-  }
-  get = getElementGetter(
-    selector,
-    async () =>
-      await $function(getTextBoxElementWithLabel, selector.label, options.selectHiddenElements),
-    inputTypesExpression,
-    options.selectHiddenElements,
-  );
-  const description = desc(selector, 'label', 'Text field');
-  return createElementWrapper(get, description, 'textBox');
+  return createElementWrapper('TextBox', 'label', attrValuePairs, _options, ...args);
 }
 
 /**
