+++ test/unit-tests/textBox.test.js
@@ -76,7 +76,7 @@ describe(test_name, () => {
 
       it('test description', async () => {
         expect(textBox('textAreaWithWrappedLabel').description).to.be.eql(
-          'Text field with label textAreaWithWrappedLabel ',
+          'TextBox with label textAreaWithWrappedLabel ',
         );
       });
     });
@@ -93,7 +93,7 @@ describe(test_name, () => {
 
       it('test description', async () => {
         expect(textBox('textAreaWithLabelFor').description).to.be.eql(
-          'Text field with label textAreaWithLabelFor ',
+          'TextBox with label textAreaWithLabelFor ',
         );
       });
     });
@@ -111,7 +111,7 @@ describe(test_name, () => {
 
       it('test description', async () => {
         expect(textBox({ id: 'textAreaWithLabelFor' }).description).to.be.eql(
-          'Text field[id="textAreaWithLabelFor"]',
+          'TextBox[id="textAreaWithLabelFor"]',
         );
       });
     });
@@ -133,7 +133,7 @@ describe(test_name, () => {
 
       it('test description', async () => {
         expect(textBox(above('textAreaWithLabelFor')).description).to.be.eql(
-          'Text field Above textAreaWithLabelFor',
+          'TextBox Above textAreaWithLabelFor',
         );
       });
     });
@@ -150,7 +150,7 @@ describe(test_name, () => {
         let elements = await textBox({
           id: 'sampleTextArea',
         }).elements();
-        expect(elements[0].description).to.be.eql('Text field[id="sampleTextArea"]');
+        expect(elements[0].description).to.be.eql('TextBox[id="sampleTextArea"]');
       });
 
       it('test value of elements', async () => {
@@ -223,7 +223,7 @@ describe(test_name, () => {
 
       it('test description', async () => {
         expect(textBox('contentEditableWithWrappedLabel').description).to.be.eql(
-          'Text field with label contentEditableWithWrappedLabel ',
+          'TextBox with label contentEditableWithWrappedLabel ',
         );
       });
     });
@@ -242,7 +242,7 @@ describe(test_name, () => {
 
       it('test description', async () => {
         expect(textBox('contentEditableWithLabelFor').description).to.be.eql(
-          'Text field with label contentEditableWithLabelFor ',
+          'TextBox with label contentEditableWithLabelFor ',
         );
       });
     });
@@ -269,7 +269,7 @@ describe(test_name, () => {
           textBox({
             id: 'contentEditableWithWrappedLabel',
           }).description,
-        ).to.be.eql('Text field[id="contentEditableWithWrappedLabel"]');
+        ).to.be.eql('TextBox[id="contentEditableWithWrappedLabel"]');
       });
     });
 
@@ -286,7 +286,7 @@ describe(test_name, () => {
 
       it('test description', async () => {
         expect(textBox(above('contentEditableWithLabelFor')).description).to.be.eql(
-          'Text field Above contentEditableWithLabelFor',
+          'TextBox Above contentEditableWithLabelFor',
         );
       });
     });
@@ -310,7 +310,7 @@ describe(test_name, () => {
         let elements = await textBox({
           id: 'sampleContentEditable',
         }).elements();
-        expect(elements[0].description).to.be.eql('Text field[id="sampleContentEditable"]');
+        expect(elements[0].description).to.be.eql('TextBox[id="sampleContentEditable"]');
       });
 
       it('test value of elements', async () => {
@@ -420,7 +420,7 @@ describe(test_name, () => {
 
         it('test description', async () => {
           expect(textBox('With Inline Text').description).to.be.eql(
-            'Text field with label With Inline Text ',
+            'TextBox with label With Inline Text ',
           );
         });
       });
@@ -437,7 +437,7 @@ describe(test_name, () => {
 
         it('test description', async () => {
           expect(textBox('With Wrapped Label').description).to.be.eql(
-            'Text field with label With Wrapped Label ',
+            'TextBox with label With Wrapped Label ',
           );
         });
       });
@@ -454,7 +454,7 @@ describe(test_name, () => {
 
         it('test description', async () => {
           expect(textBox('With Label For').description).to.be.eql(
-            'Text field with label With Label For ',
+            'TextBox with label With Label For ',
           );
         });
       });
@@ -481,7 +481,7 @@ describe(test_name, () => {
             textBox({
               id: inputType.name + 'WithLabelFor',
             }).description,
-          ).to.be.eql(`Text field[id="inputType-${inputType.type}WithLabelFor"]`);
+          ).to.be.eql(`TextBox[id="inputType-${inputType.type}WithLabelFor"]`);
         });
 
         it('should return false for hidden element when isVisible fn is called', async () => {
@@ -507,7 +507,7 @@ describe(test_name, () => {
 
         it('test description', async () => {
           expect(textBox(above('With Label For')).description).to.be.eql(
-            'Text field Above With Label For',
+            'TextBox Above With Label For',
           );
         });
       });
@@ -524,7 +524,7 @@ describe(test_name, () => {
           let elements = await textBox({
             id: `sample${inputType.type}`,
           }).elements();
-          expect(elements[0].description).to.be.eql(`Text field[id="sample${inputType.type}"]`);
+          expect(elements[0].description).to.be.eql(`TextBox[id="sample${inputType.type}"]`);
         });
 
         it('test value of elements', async () => {
@@ -590,7 +590,7 @@ describe(test_name, () => {
 
       it('test description', async () => {
         expect(textBox('With Inline Text').description).to.be.eql(
-          'Text field with label With Inline Text ',
+          'TextBox with label With Inline Text ',
         );
       });
     });
@@ -607,7 +607,7 @@ describe(test_name, () => {
 
       it('test description', async () => {
         expect(textBox('With Wrapped Label').description).to.be.eql(
-          'Text field with label With Wrapped Label ',
+          'TextBox with label With Wrapped Label ',
         );
       });
     });
@@ -624,7 +624,7 @@ describe(test_name, () => {
 
       it('test description', async () => {
         expect(textBox('With Label For').description).to.be.eql(
-          'Text field with label With Label For ',
+          'TextBox with label With Label For ',
         );
       });
     });
@@ -651,7 +651,7 @@ describe(test_name, () => {
           textBox({
             id: inputTypeName + 'WithLabelFor',
           }).description,
-        ).to.be.eql('Text field[id="input-without-typeWithLabelFor"]');
+        ).to.be.eql('TextBox[id="input-without-typeWithLabelFor"]');
       });
     });
 
@@ -666,7 +666,7 @@ describe(test_name, () => {
 
       it('test description', async () => {
         expect(textBox(above('With Label For')).description).to.be.eql(
-          'Text field Above With Label For',
+          'TextBox Above With Label For',
         );
       });
 
@@ -683,7 +683,7 @@ describe(test_name, () => {
 
       it('test description of elements', async () => {
         let elements = await textBox('sampleInputWithoutType').elements();
-        expect(elements[0].description).to.be.eql('Text field with label sampleInputWithoutType ');
+        expect(elements[0].description).to.be.eql('TextBox with label sampleInputWithoutType ');
       });
 
       it('test value of elements', async () => {
